Separate line-splitter function into its own script
