+++ scripts/text_formatter.py
@@ -1,5 +1,3 @@
-import csv
-import os
 import re
 import unicodedata as ud
